backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
authlib==1.2.1
httpx>=0.24.0
minio==7.2.0
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0
python-dotenv==1.0.0
Pillow>=10.0.0
SimpleITK>=2.3.0
numpy>=1.24.0

